+++ docs/guide/source/changelog.rst
@@ -25,7 +25,7 @@ Unreleased (MuJoCo 3.3.7)
 
 - Changes to :docs-rs:`~mujoco_rs::viewer::<struct>MjViewer`:
 
-  - Added :docs-rs:`~~mujoco_rs::viewer::<struct>MjViewer::<method>add_custom_ui_widgets` for custom UI widgets.
+  - Added :docs-rs:`~~mujoco_rs::viewer::<struct>MjViewer::<method>add_ui_callback` for custom UI widgets.
     This allows users to create custom windows, panels, and other UI elements using `egui <https://github.com/emilk/egui>`_.
     See :ref:`custom_ui_widgets` for more information.